requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
fastapi==0.109.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.31.2
idna==3.11
pdf2image==1.16.3
pillow==10.2.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyMuPDF==1.26.7
pyparsing==3.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.5
rsa==4.9.1
starlette==0.35.1
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.0
azure-ai-formrecognizer==3.3.0
openai==1.12.0
anthropic>=0.39.0
langgraph>=0.2.0
langchain-core>=0.3.0
